move datasource into schema folder
